Fixed build on Windows (#10)
* Fixed build on Windows

* Downloading conan.cmake from git

* Updated Vulkan version on linux

* Conan profile detect

* Conan options fix

* Conan options fix

* Conan options fix

* Revert "Updated Vulkan version on linux"

This reverts commit df36c3bfab15c9a01d80bd834f730e408730b7b8.

* Hardcoded version of conan

* Upgraded vulkan linux version

* Added apt dependencies for linux

* Fixed conan version for Windows

* Use g++11 on linux

* Fixed workflow error

* Fixed build error

* Removed sudo in windows

* Wno-error=unknown-warning

* Fixed GCC compile options

* Removed Clang specific compile option from GCC

* Fixed IsZero bug

* Fixed switch case bug

* Fixed cast error

* Fixed cast error again

* Fixed compile error

* Removed light moving

* Fixed compile errors

* Added -fno-semantic-interposition

* Tried to fix nullptr in logger

* Removed -Wnull-dereference

* Fixed exception order

* Fixed window type on linux

* Fixed window type on linux

* Fixed compile error

* Fixed compile error

* Updated way of packaging

* Fixed build on Windows

* Update cmake.yml

* Update cmake.yml

+++ Sources/Vulkan/VulkanImage.cpp
@@ -408,7 +408,7 @@ VulkanImage::GenerateMipmaps(
                 };
                 auto dstOffsets = std::array<vk::Offset3D, 2> {
                     vk::Offset3D(0, 0, 0),
-                    vk::Offset3D(mipWidth > 1 ? mipWidth / 2 : 1, mipHeight > 1 ? mipHeight / 2 : 1, 1)
+                    vk::Offset3D(static_cast<std::int32_t>(mipWidth > 1 ? mipWidth / 2 : 1), static_cast<std::int32_t>(mipHeight > 1 ? mipHeight / 2 : 1), 1)
                 };
 
                 auto blit = vk::ImageBlit()
